make sure we return doubles instead of ints

+++ src/edge_probability.cpp
@@ -51,7 +51,7 @@ namespace mjd {
 using namespace Rcpp;
 // for testing we will wrap and export this function so it is available in R.
 // [[Rcpp::export]]
-int ep(arma::vec intercepts,
+double ep(arma::vec intercepts,
        arma::mat coefficients,
        NumericVector latent_pos,
        int sender,
